+++ wagtail/wagtailadmin/api/urls.py
@@ -3,8 +3,11 @@ from __future__ import absolute_import, unicode_literals
 from django.conf.urls import url
 
 from wagtail.api.v2.router import WagtailAPIRouter
+from wagtail.wagtaildocs.api.admin.endpoints import DocumentsAdminAPIEndpoint
+from wagtail.wagtailimages.api.admin.endpoints import ImagesAdminAPIEndpoint
 
-from .endpoints import DocumentsAdminAPIEndpoint, ImagesAdminAPIEndpoint, PagesAdminAPIEndpoint
+from .endpoints import PagesAdminAPIEndpoint
+
 
 v1 = WagtailAPIRouter('wagtailadmin_api_v1')
 v1.register_endpoint('pages', PagesAdminAPIEndpoint)
